feat(algorithms): changed the way of how algorithms set the pointers to parent (boarding) stops; now there is an option to include (or not) intermediate legs when reconstructing the journeys
Before, both Basic and WMC variants set, for each label, the pointer (boarding_stop) to each intermediate stop of a given leg (part of a journey done on the same trip), as opposed to pointing just to the first stop of the leg. Notably, this meant that journey reconstruction inherently included intermediate stops. Now, that functionality is only handled in the reconstruction process via the new boolean flag.